+++ src/main/java/com/kitchenhack/apikitchen/repositories/ProgresoSaludRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface ProgresoSaludRepository extends JpaRepository<ProgresoSalud, Integer> {
-    Optional<ProgresoSalud> findByIdUsuario_Id(Integer usuarioId);
+    Optional<ProgresoSalud> findByIdUsuario_Id(Long usuarioId);
 }
